vm: rename vmstate/eei
vm: move eei/vmstate into new EEI dir
